Added timeout on sync event and added ability to exit program with M81

+++ redeem/gcodes/M81.py
@@ -15,10 +15,15 @@ import os
 class M81(GCodeCommand):
 
     def execute(self, g):
-        os.system("shutdown -h now")
+        if g.has_letter("P"):         
+            g.answer = None   # Prevent reply
+            self.printer.redeem.running = False
+            self.printer.path_planner.queue_sync_event(True)
+        else:
+            os.system("shutdown -h now")
 
     def get_description(self):
-        return "Shutdown the whole Replicape controller board"
+        return "Shutdown the whole Replicape controller board. If paramter P is present, only exit loop. "
 
     def is_buffered(self):
         return False
